Rename directory, properly organize files
Aligning directory and logical structure to more closely match other
mods, e.g., anegostudios' mods. Some code patching system developments
were made in preparation for the next commit.

Setup is complete to refactor the patches. Will be moving whatever
possible to JSON patches to clarify codebase. The final goal of this
branch is to introduce a flexible horticulture system to be iterated
upon in the future. It will do so by introducing genetic behaviors that
are assumed to be coupled with the horticulture system, providing a
clearer path to decouple genetics from the system in the future.
