+++ Origins/BlockBehavior/BehaviorShowBEBlockInfo.cs
@@ -1,4 +1,5 @@
 ﻿using Newtonsoft.Json.Linq;
+using Origins.Util;
 using System.Linq;
 using System.Text;
 using Vintagestory.API.Common;
@@ -7,7 +8,7 @@ using Vintagestory.API.MathTools;
 using Vintagestory.API.Util;
 using Vintagestory.GameContent;
 
-namespace Origins.Patches.Behaviors;
+namespace Origins.GameContent;
 
 internal class BlockBehaviorFarmlandGeneticData : BlockBehavior, ICodePatch
 {
