+++ Origins/Item/ItemSeedAnalyzer.cs
@@ -1,0 +1,44 @@
+﻿using Vintagestory.API.Common;
+
+namespace Origins.GameContent;
+
+internal class ItemSeedAnalyzer : Item
+{
+    bool hortsysEnabled = false;
+    HorticultureSystem HortSys;
+
+    public override void OnLoaded(ICoreAPI api)
+    {
+        base.OnLoaded(api);
+
+        hortsysEnabled = api.ModLoader.IsModSystemEnabled("Origins.Systems.Horticulture.HorticultureSystem");
+
+    }
+
+    public override void OnHeldInteractStart(ItemSlot slot, EntityAgent byEntity, BlockSelection blockSel, EntitySelection entitySel, bool firstEvent, ref EnumHandHandling handling)
+    {
+        handling = EnumHandHandling.NotHandled;
+
+        // TODO (chris): CONFIG OPTION; add setting to change between "Toggle" and "Continuous" behavior
+        if (!HortSysLoaded())
+        {
+            api.Logger.Error("Oops, HorticultureSystem was not found! You should not see this.");
+            return;
+        }
+
+#pragma warning disable CS0618 // not working on this yet
+        HortSys
+            .GetHudSeedAnalyzer()
+            .WithBlockAndEntitySelection(blockSel, entitySel)
+            .Toggle();
+#pragma warning restore CS0618 // Type or member is obsolete
+    }
+
+    private bool HortSysLoaded()
+    {
+        if (hortsysEnabled && HortSys == null)
+            HortSys = api.ModLoader.GetModSystem<HorticultureSystem>();
+
+        return HortSys != null;
+    }
+}
