+++ Origins/assets/origins/config/paths.json
