+++ Origins/assets/origins/itemtypes/seedanalyzer.json
