+++ Origins/assets/origins/textures/item/magicwand.png
