--- origins/Patches/IPatch.cs
@@ -1,14 +1,0 @@
-using Vintagestory.API.Common;
-
-namespace Origins.Patches
-{
-    interface IPatch
-    {
-        public static abstract void RegisterPatch(ICoreAPI api);
-    }
-
-    interface ICodePatch : IPatch
-    {
-        public static abstract void ApplyPatch(ICoreAPI api);
-    }
-}
